refactor: move `Button` variant constants to `variants.ts` file
