+++ src/components/ui/button/index.ts
@@ -1,9 +1,5 @@
 export {
 	Button,
-	BUTTON_VARIANTS,
-	BUTTON_TEXT_SIZES,
-	BUTTON_ICON_SIZES,
-	BUTTON_SIZES,
 } from "./button";
 
-export { buttonVariants } from "./variants";
+export { buttonVariants, BUTTON_VARIANTS, BUTTON_TEXT_SIZES, BUTTON_ICON_SIZES, BUTTON_SIZES } from "./variants";
